third_party/WebKit/LayoutTests/http/tests/htmlimports/imported-document.html

<!DOCTYPE html>
<html>
<head>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<link id="target" rel="import" href="resources/hello.html">
</head>
<body>
<script>
test(function() {
  var target = document.getElementById("target");
  assert_equals(target.href, target.import.URL);
  assert_true(target.import instanceof Document)
},
"Check on the document URL of loaded document",
{
  "help": ["https://dvcs.w3.org/hg/webcomponents/raw-file/tip/spec/imports/index.html#imported-document"],
});
</script>
</body>
</html>
